move numeric interval constraints to numeric.py
